+++ src/DataJournal.ts
