+++ src/pdw.ts
@@ -1,120 +1,8 @@
 import { Temporal } from "temporal-polyfill";
+import * as dj from "./DataJournal";
+import { Period } from "./Period";
 
 //#region ### TYPES ###
-
-/**
- * A alias for string, implying one with the structure of:
- * {@link EpochStr}-{@link SmallID}
- */
-export type UID = string;
-
-/**
- * A alias for string, a string of 4 random characters
- * generated by the {@link makeSmallID} function
- */
-export type SmallID = string;
-
-/**
- * An alias for string, a string that's made from the number of 
- * milliseconds that have occurred since the epoch, stored in base 36
- * A short, consistenty way to track timezone-dependent fields like
- * {@link _created} and {@link _updated} that isn't 
- * subject to alternative interpretations by, say, Excel
- */
-export type EpochStr = string;
-
-/**
- * An alias for string, a string that can be parsed by
- * Temporal.Duration.from()
- * Example: 
- * `P3D6H30M`
- */
-export type DurationStr = string;
-
-/**
- * A alias for string, any valid Period.toString() string
- */
-export type PeriodStr = string;
-
-/**
- * An alias for string, intended to be markdown-enabled in use
- */
-export type Markdown = string
-
-/**
- * This "Omit" type is a mechanism to copy types/interfaces, *except* the given keys
- * Example use: Omit<EntryLike, '_did'>;
- */
-type Omit<T, K extends keyof T> = {
-    [P in keyof T as P extends K ? never : P]: T[P];
-};
-
-//#endregion
-
-//#region ### ENUMS ###
-
-export enum PointType {
-    /**
-     * a number
-     */
-    NUMBER = "NUMBER",
-    /**
-     * A string, assumed to be short-ish
-     */
-    TEXT = 'TEXT',
-    /**
-     * A string, assumed to be long-ish, to run through 'marked' if possible
-     */
-    MARKDOWN = 'MARKDOWN',
-    /**
-     * Essentially an enumeration
-     */
-    SELECT = 'SELECT',
-    /**
-     * true or false
-     */
-    BOOL = 'BOOL', //true false
-    /**
-     * A Temporal Duration string, starting with the 'P' (and usually also 'T') to really ensure no ambiguity
-     * P3M = 3 Months
-     * PT3M = 3 Minutes
-     */
-    DURATION = 'DURATION',
-    /**
-     * A Temporal PlainTime string (no timezone)
-     */
-    TIME = 'TIME', //Temporal.plainTime
-    /**
-     * An array of _tid
-     */
-    MULTISELECT = 'MULTISELECT', //Comma-separated list of _tid 
-}
-
-/**
- * Level of granularity. Assigned to each {@link Def}
- */
-export enum Scope {
-    SECOND = 'SECOND',
-    MINUTE = 'MINUTE',
-    HOUR = 'HOUR',
-    DAY = 'DAY',
-    WEEK = 'WEEK',
-    MONTH = 'MONTH',
-    QUARTER = 'QUARTER',
-    YEAR = 'YEAR',
-}
-
-/**
- * Default operation to apply when performing a {@link PointRollup}. 
- * Each {@link PointDef} is associated with one Rollup.
- */
-export enum Rollup {
-    COUNT = 'COUNT',
-    COUNTUNIQUE = 'COUNTUNIQUE',
-    SUM = 'SUM',
-    AVERAGE = 'AVERAGE',
-    COUNTOFEACH = 'COUNTOFEACH',
-}
 
 //#endregion
 
@@ -170,192 +58,6 @@ export interface CanonicalDataTranslator {
      */
     // exportTo(allData: CanonicalDataset, params: any): any
 }
-/**
- * Basic data filtering parameters, supported by the {@link PDW} methods for
- * {@link getDefs} & the 2 other element "getters". The {@link DataStore} methods
- * will get {@link ReducedParams} passed to them by the PDW methods, which will
- * perform the sanitization.
- * Not all parameters are considered for each getter, but it should make
- * things a bit simpler to standardize the params
- */
-export interface StandardParams {
-    /**
-     * Include things marked as deleted?
-     * no - default
-     * yes - include all
-     * only - only include deleted things
-     */
-    includeDeleted?: 'yes' | 'no' | 'only',
-    /**
-     * Lower-bound of Entry.period.
-     * Like all Periods, does not contain Time Zone info.
-     * Will be treated as the FIRST SECOND of the Period or PeriodStr provided.
-     * Example:
-     * "2023-Q2" would be synonymous with "2023-04-01T:00:00:00"
-     */
-    from?: Period | PeriodStr,
-    /**
-     * Upper-bound of Entry.period. 
-     * Like all Periods, does not contain Time Zone info.
-     * Will be treated as the LAST SECOND of the Period or PeriodStr provided.
-     * Example:
-     * "2023-Q2" would be synonymous with "2023-08-31T:23:59:59"
-     */
-    to?: Period | PeriodStr,
-    /**
-     * Entry period. Sets the Query.from and Query.to values internally.
-     */
-    inPeriod?: Period | string
-    /**
-     * The lower-bound of Element.created, represented as an {@link EpochStr}
-     * or Temporal.ZonedDateTime. 
-     */
-    createdAfter?: Temporal.ZonedDateTime | EpochStr,
-    /**
-     * The upper-bound of Element.created, represented as an {@link EpochStr}
-     * or Temporal.ZonedDateTime. 
-     */
-    createdBefore?: Temporal.ZonedDateTime | EpochStr,
-    /**
-     * The lower-bound of Element.updated, represented as an {@link EpochStr}
-     * or Temporal.ZonedDateTime. 
-     */
-    updatedAfter?: Temporal.ZonedDateTime | EpochStr,
-    /**
-     * The upper-bound of Element.updated, represented as an {@link EpochStr}
-     * or Temporal.ZonedDateTime. 
-     */
-    updatedBefore?: Temporal.ZonedDateTime | EpochStr,
-    /**
-     * A list of Element.uid. Will filter to Elements in the list.
-     */
-    uid?: UID[] | UID,
-    /**
-     * A list of {@link SmallID}.
-     * For Entry and Def elements, will return those with Element.did in the list.
-     */
-    did?: SmallID[] | SmallID,
-    /**
-     * A list of {@link UID}.
-     * Will return any Entry instance(s) whose Entry.eid is in the list.
-     */
-    eid?: UID[] | UID,
-    /**
-     * A list of strings for the associated with Def.lbl
-     * Internally, this is translated into a list of _did strings, and behaves like
-     * {@link StandardParams.did}
-     * For Entry and Def elements, will return those with Element.did in the list.
-     */
-    defLbl?: string[] | string,
-    //pointLbl?: string[] | string, do you want to support this
-    /**
-     * A list of strings.
-     * Will return any Tag(s) whose Tag.lbl is in the list.
-     */
-    tag?: string,
-    /**
-     * Reduces the resulting {@link Def} and {@link} Entry results to those whose
-     * scope is in the provided list.
-     */
-    scope?: Scope | Scope[];
-    /**
-     * A limit on the number of responses returned. Probably unsorted and therefore
-     * not super helpful, but this is something I'd like to support.
-     */
-    limit?: number,//#TODO
-    /**
-     * If an empty query is sent without this field, the query is rejected.
-     * This is prevent the error trap of accidentally asking for *everything*,
-     * which could be expensive in cloud-based datastores.
-     */
-    allOnPurpose?: boolean
-    /**
-     * If provdided, will query for the current day. Will ignore any provided values
-     */
-    today?: any;
-    /**
-     * If provdided, will query for the current week. Will ignore any provided values
-     */
-    thisWeek?: any;
-    /**
-     * If provdided, will query for the current month. Will ignore any provided values
-     */
-    thisMonth?: any;
-    /**
-     * If provdided, will query for the current quarter. Will ignore any provided values
-     */
-    thisQuarter?: any;
-    /**
-     * If provdided, will query for the current year. Will ignore any provided values
-     */
-    thisYear?: any;
-    /**
-     * If provided, the query response will include Period rollups
-     */
-    rollup?: Scope
-}
-
-/**
- * A more tightly-controlled version of the {@link StandardParams}, meant to make
- * development of any new {@link DataStore} **EASIER** to implement. SanitizedParams are
- * what DataStore versions of the main setter/getter functions respond to.  
- */
-export interface ReducedParams {
-    /**
-     * yes - excludes deleted stuff
-     * no - returns deleted & undeleted stuff
-     * only - returns **only** deleted stuff
-     */
-    includeDeleted: 'yes' | 'no' | 'only',
-    from?: Period,
-    to?: Period,
-
-    //these 4 pairs come in pairs from the sanitizeParams() func, DataStore could use either
-    createdAfter?: Temporal.ZonedDateTime,
-    createdAfterEpochStr?: EpochStr,
-
-    createdBefore?: Temporal.ZonedDateTime,
-    createdBeforeEpochStr?: EpochStr,
-
-    updatedBefore?: Temporal.ZonedDateTime
-    updatedBeforeEpochStr?: EpochStr,
-
-    updatedAfter?: Temporal.ZonedDateTime,
-    updatedAfterEpochStr?: EpochStr,
-
-    uid?: UID[],
-    did?: SmallID[],
-    eid?: UID[],
-    defLbl?: string[],
-    pointLbl?: string[],
-
-    limit?: number, //#TODO
-    /**
-     * Reduces the resulting {@link Def} and {@link} Entry results to those whose
-     * scope is in the provided list.
-     */
-    scope?: Scope[];
-    allOnPurpose?: boolean
-}
-
-export interface ElementDataMap {
-    defs: DefData[];
-    entries: EntryData[];
-}
-
-export interface ElementMap {
-    defs: Def[];
-    entries: Entry[];
-}
-
-/**
-    The main data structure.  
-    Contanis a map of DefLike and EntryLike objects, along 
-    with an optional overview.
- */
-export interface CanonicalDataset extends ElementDataMap {
-    overview?: DataStoreOverview;
-}
 
 export interface Transaction {
     create: ElementMap;
@@ -363,207 +65,11 @@ export interface Transaction {
     delete: DeletionMsgMap;
 }
 
-export interface DeletionMsg {
-    uid: UID;
-    deleted: boolean;
-    updated: EpochStr;
-}
-
-export interface DeletionMsgMap {
-    defs: DeletionMsg[];
-    entries: DeletionMsg[];
-}
-
-/**
- * This interface is extended by the interfaces for the base Elements
- */
-export interface ElementLike {
-    /**
-     * Definition ID - required
-     */
-    _did: string;
-    /**
-     * Universal ID - uniquely identifies an INSTANCE of any element.
-     */
-    _uid?: UID;
-    /**
-     * Was the element marked as deleted?
-     * For things like "_did" or "_pid", "_tid", you may have multiple
-     * instances with the same value, but you'll tell them apart using
-     * the _uid --- and only **one** will have _deleted == false
-     */
-    _deleted?: boolean;
-    /**
-     * When the element was created
-     */
-    _created?: EpochStr;
-    /**
-     * When the element was updated, usually lines up with "_created"
-     * unless the instance of the element was created via updating a 
-     * previous instance
-     */
-    _updated?: EpochStr;
-    /**
-     * other key/value pairs will attempt to find
-     * the PointDef who's _pid == the key and spin
-     * up a new {@link EntryPoint} who's _val is 
-     * the value
-     */
-    [x: string]: any;
-}
-
-/**
- * Definitions Data Shape
- */
-export interface DefLike extends ElementLike {
-
-    /**
-     * The label for the Definition
-     */
-    _lbl?: string;
-    /**
-     * A text string describing what the data are
-     */
-    _desc?: string;
-    /**
-     * A shorthand for these types of data
-     */
-    _emoji?: string;
-    /**
-     * Scope of the definition (e.g. "day", "week"...)
-     */
-    _scope?: Scope;
-    /**
-     * For things you want to collect, but not necessarily see
-     */
-    _tags?: string[];
-    /**
-     * The points on the definition
-     */
-    _pts?: PointDefLike[];
-}
-
-export interface PointDefLike {
-    /**
-     * Point Definition ID.
-     */
-    _pid?: SmallID
-    /**
-    * Point type
-    */
-    _type?: PointType;
-    /**
-     * Default rollup type
-     */
-    _rollup?: Rollup;
-
-    /**
-     * The label for the Definition or PointDef
-     */
-    _lbl?: string;
-    /**
-     * A text string describing what the data are
-     */
-    _desc?: string;
-    /**
-     * A shorthand for these types of data
-     */
-    _emoji?: string;
-    /**
-     * List of potential options for Select & Multiselect
-     */
-    _opts?: string[];
-}
-
-export interface EntryLike extends ElementLike {
-    /**
-     * Entry ID, a unique identifier for an entry - unlike {@link _uid} 
-     * _eid is not updated when an entry is updated. A new _uid is created
-     * for the updated Entry, but they will share _eid values.
-     */
-    _eid?: UID,
-    /**
-     * A generic note about an entry, all entries can have them
-     */
-    _note?: string,
-    /**
-     * When the entry is for
-     */
-    _period?: PeriodStr,
-    /**
-     * For tracking where the tracking is coming from
-     */
-    _source?: string,
-
-    /**
-     * _pid keys, with the value of the EntryPoint
-     */
-    [_pid: string]: any
-}
-
-export interface ElementData {
-    _uid: UID;
-    _deleted: boolean;
-    _created: EpochStr;
-    _updated: EpochStr;
-    _did: string;
-    [x: string]: any;
-}
-
-export interface DefData extends ElementData {
-    _lbl: string;
-    _desc: string;
-    _emoji: string;
-    _tags: string[];
-    _scope: Scope;
-    _pts: PointDefData[];
-}
-
-export interface PointDefData {
-    _pid: SmallID
-    _lbl: string;
-    _desc: string;
-    _emoji: string;
-    _type: PointType;
-    _rollup: Rollup;
-    _opts?: string[];
-}
-
-export interface EntryData extends ElementData {
-    _eid: UID,
-    _note: string,
-    _period: PeriodStr,
-    _source: string,
-    [_pid: string]: any
-}
-
-export interface PointRollup {
-    pid: SmallID;
-    lbl: string;
-    emoji: string;
-    val: any;
-    method: Rollup;
-    vals: any[];
-}
-
-export interface EntryRollup {
-    did: SmallID;
-    lbl: string;
-    emoji: string;
-    pts: PointRollup[];
-}
 
 export interface PeriodSummary {
     period: PeriodStr | "ALL";
-    entryRollups: EntryRollup[];
+    // entryRollups: EntryRollup[];
     entries: EntryData[]
-}
-
-/**
- * Option object, map with key = _oid, val = text label for option
- */
-export interface OptMap {
-    [_oid: SmallID]: string;
 }
 
 export interface QueryResponse {
@@ -615,7 +121,7 @@ export interface DataStoreOverview {
 
 export class PDW {
     dataStore: DataStore;
-    private _manifest: Def[];
+    private _manifest: Def[]; //#TODO - move to Database Class
     private constructor(manifest: DefLike[], store?: DataStore,) {
         this._manifest = this._manifest = manifest.map(defData => new Def(defData, this));
         if (store !== undefined) {
@@ -1248,1227 +754,6 @@ export class PDW {
             })
             return entryTypes
         }
-    }
-}
-
-/**
- * Base class 
- */
-abstract class Element {
-    data: ElementData;
-    pdw: PDW;
-    __modified: boolean;
-    __deletionChange?: boolean;
-    readonly tempCreated: Temporal.ZonedDateTime;
-    readonly tempUpdated: Temporal.ZonedDateTime;
-
-    constructor(inputData: ElementLike, pdwRef: PDW) {
-        this.pdw = pdwRef;
-        this.data = {
-            _did: inputData._did,
-            _uid: inputData._uid ?? makeUID(),
-            _deleted: Element.handleDeletedInputVariability(inputData._deleted),
-            _created: Element.handleEpochStrInputVariability(inputData._created),
-            _updated: Element.handleEpochStrInputVariability(inputData._updated)
-        }
-        this.tempCreated = parseTemporalFromEpochStr(this.data._created);
-        this.tempUpdated = parseTemporalFromEpochStr(this.data._updated);
-        this.__modified = false;
-    }
-
-    /**
-     * The identifier for the Definition associated with this element.
-     */
-    get did() {
-        return this.data._did;
-    }
-
-    /**
-     * The unique identifier of the Element instance. Should not have any duplicates 
-     * anywhere in the PDW.
-     */
-    get uid() {
-        return this.data._uid;
-    }
-    /**
-     * An {@link EpochStr} describing when the Element was created.
-     * If an Element is updated, the _created property will be carried forward.
-     */
-    get created(): string {
-        return this.data._created as string;
-    }
-    /**
-     * An {@link EpochStr} describing when the Element was most recently updated.
-     * If an Element is updated, the _updated property will be set to the moment that Element
-     * instance was created. The _updated property will also be set to the time the Element
-     * class instance was marked as Deleted (or Undeleted).
-     */
-    get updated(): string {
-        return this.data._updated as string;
-    }
-    /**
-    * If an Element should be considered deleted, because things aren't actually *deleted* in 
-    * order to make it possible to combine datasets and maintain integrity.
-    */
-    get deleted() {
-        return this.data._deleted;
-    }
-    set created(newDate: EpochStr | Temporal.ZonedDateTime | Temporal.Instant | Date) {
-        this.data._created = Element.handleEpochStrInputVariability(newDate);
-        this.__modified = true;
-    }
-    set updated(newDate: EpochStr | Temporal.ZonedDateTime | Temporal.Instant | Date) {
-        this.data._updated = Element.handleEpochStrInputVariability(newDate);
-        this.__modified = true;
-    }
-    set deleted(isDeleted: boolean) {
-        this.data._deleted = isDeleted;
-        this.__modified = true;
-        this.__deletionChange = isDeleted;
-    }
-
-    /**
-     * Create a **static** copy of the Element's ElementData.
-     */
-    toData(): ElementData | DefData | EntryData {
-        return JSON.parse(JSON.stringify(this.data));
-    }
-
-    /**
-     * Marks the existing Element as deleted and sends a copy of it with
-     * its current property values and a new _uid to the connected {@link DataStore}(s)
-     * @returns this
-     */
-    async save() {
-        if (this.__modified === false) {
-            console.warn('Tried saving an unchanged Element, ignoring request');
-            return;
-        }
-        const oldId = this.data._uid;
-        const changeTime = makeEpochStr();
-
-        let created: ElementLike[] = [];
-        let deleted: DeletionMsg[] = [];
-
-        //deletions and un-deletions need to be handled differently, since they don't spawn 
-        //new elements to be created. Right now I'm handling it this way.
-        if (this.__deletionChange !== undefined) { //indicates 'delete' was changed.
-            this.data._updated = changeTime;
-            deleted.push({
-                uid: this.data._uid,
-                deleted: this.data._deleted,
-                updated: changeTime
-            });
-            this.__deletionChange = undefined;
-
-        } else {
-            //generate a deletion message for stores
-            deleted.push({
-                uid: oldId,
-                deleted: true,
-                updated: changeTime
-            });
-            this.data._uid = makeUID();
-            this.data._updated = changeTime;
-            created.push(this.toData());
-        }
-
-        //return the 'modified' flag to false to indicate it's not been deleted
-        this.__modified = false;
-
-        const elementType = this.getType();
-        if (elementType === 'DefData') {
-            await this.pdw.setDefs(created, [], deleted);
-            return this
-        }
-        if (elementType === 'EntryData') {
-            await this.pdw.setEntries(created as EntryData[], [], deleted)
-            return this
-        }
-        throw new Error('What kind of element is this anyway? ' + elementType)
-    }
-
-    isSaved() {
-        return !this.__modified;
-    }
-
-    setUpdatedToNow() {
-        this.data._updated = makeEpochStr();
-    }
-
-    /**
-     * Checks if the argument was updated more recently than this
-     * @param elementData ElementData data to compare against
-     * @returns true if argument is updated more recently than this
-     */
-    isOlderThan(elementData: ElementData | Element) {
-        if (elementData.hasOwnProperty('__modified')) return this.data._updated < elementData.updated;
-        if (elementData.hasOwnProperty('_updated')) return this.data._updated < (<ElementData>elementData)._updated;
-        throw new Error('What props does it have?')
-        //ran into a situation one time where I saved something twice on successive
-        //lines nad only the first one was taken due to the "strictly olderthan",
-        //but I don't want mass overwrites of the same data, so I'm living with
-        //that for now.
-    }
-
-    /**
-     * Checks whether this is the same Element as the comparison AND if the 
-     * comparison is newer than this
-     * @param comparison ElementData data that might be a newer copy of this
-     * @returns true if the argument is a newer version of the same Element
-     */
-    shouldBeReplacedWith(comparison: ElementData | Element): boolean {
-        return Element.shouldReplace(comparison, this.data)
-    }
-
-    /**
-     * Get the type of an Element.
-     * @returns string representing the type of element
-     */
-    getType(): 'DefData' | 'EntryData' | null {
-        return Element.getTypeOfElement(this.data)!
-    }
-
-    /**
-     * Checks to see if this has the same _did (or _eid, _tid, _pid) as
-     * whatever is passed in
-     * @param comparison Element to compare against
-     */
-    sameIdAs(comparison: ElementData | Element) {
-        if (comparison.hasOwnProperty('__modified')) return Element.hasSameId(this.data, comparison.data);
-        return Element.hasSameId(this.data, (<ElementData>comparison));
-    }
-
-    sameTypeAs(comparison: ElementData | Element) {
-        if (comparison.hasOwnProperty('__modified')) return this.getType() === Element.getTypeOfElement(comparison.data);
-        return this.getType() === Element.getTypeOfElement((<ElementData>comparison));
-    }
-
-    passesFilters(params: ReducedParams) {
-        if (params.uid !== undefined && !params.uid.some(uid => uid === this.data._uid)) return false;
-        if (params.did !== undefined && this.data._did !== undefined && !params.did.some(did => did === this.data._did)) return false;
-        if (params.eid !== undefined && this.data._eid !== undefined && !params.eid.some(eid => eid === this.data._eid)) return false;
-        // if (params.pid !== undefined && this.data._pid !== undefined && !params.pid.some(pid => pid === this.data._pid)) return false;
-
-        const type = this.getType()!;
-
-        if (params.defLbl !== undefined && type === 'DefData' && !params.defLbl.some(lbl => lbl === this.data._lbl)) return false;
-
-        if (params.createdBeforeEpochStr !== undefined && params.createdBeforeEpochStr < this.data._created) return false;
-        if (params.createdAfterEpochStr !== undefined && params.createdAfterEpochStr > this.data._created) return false;
-
-        if (params.updatedBeforeEpochStr !== undefined && params.updatedBeforeEpochStr < this.data._updated) return false;
-        if (params.updatedAfterEpochStr !== undefined && params.updatedAfterEpochStr > this.data._updated) return false;
-
-        if (params.includeDeleted === 'no' && this.data._deleted === true) return false;
-        if (params.includeDeleted === 'only' && this.data._deleted === false) return false;
-
-        if (type === 'EntryData') {
-            if (params.from !== undefined && (<Entry><unknown>this).period.getStart().isBefore(params.from)) return false
-            if (params.to !== undefined && (<Entry><unknown>this).period.getEnd().isAfter(params.to)) return false
-        }
-
-        return true;
-    }
-
-    /**
-     * Make a de-referenced COPY of this Element. 
-     * NOTE: Not the Element's *data*, that would be the {@link toData} method.
-     * @returns an instance of {@link Tag}, {@link Def}, or {@link Entry} that is a de-referenced copy of this
-     */
-    makeStaticCopy(): Def | Entry {
-        const type = Element.getTypeOfElement(this.data);
-        const data = this.toData();
-        if (type === 'DefData') return new Def(data, this.pdw);
-        if (type === 'EntryData') return new Entry(data, this.pdw);
-        throw new Error('What type was this? ' + type);
-    }
-
-    // extraction & "never nesting" really made this code way more readable... huh
-    private static handleDeletedInputVariability(deletedVal: any): boolean {
-        if (deletedVal === undefined) return false
-        if (typeof deletedVal === 'boolean') return deletedVal;
-        if (typeof deletedVal === 'string') return (<string>deletedVal).toUpperCase() === 'TRUE';
-        if (typeof deletedVal === 'number') return (<number>deletedVal) === 1 ? true : false
-        console.warn(`Didn't know how to set '_deleted' based on input. Defaulting to 'false':`, deletedVal);
-        return false
-    }
-
-    private static handleEpochStrInputVariability(inputSeen: any): EpochStr {
-        if (inputSeen === undefined) return makeEpochStr();
-        if (typeof inputSeen === 'string') {
-            inputSeen = inputSeen.trim(); //added for purpose at one point
-            if (isValidEpochStr(inputSeen)) return inputSeen;
-            //try passing through new Date()'s wide-open interpretations
-            return makeEpochStrFrom(Temporal.Instant.fromEpochMilliseconds(new Date(inputSeen).getTime()).toZonedDateTimeISO(Temporal.Now.timeZone()))!;
-        }
-        if (typeof inputSeen === 'number') {
-            makeEpochStrFrom(Temporal.Instant.fromEpochMilliseconds(inputSeen).toZonedDateTimeISO(Temporal.Now.timeZone()));
-        }
-        console.warn('Did not know how to handle parsing this date input, defaulting to now:', inputSeen);
-        return makeEpochStr();
-    }
-
-    static shouldReplace(thing: ElementData | Element, thingItMayReplace: ElementData | Element): boolean {
-        if (thing.hasOwnProperty('__modified')) thing = thing.data
-        if (thingItMayReplace.hasOwnProperty('__modified')) thingItMayReplace = thingItMayReplace.data
-
-        if (!Element.hasSameId((<ElementData>thing), (<ElementData>thingItMayReplace))) return false;
-        if ((<ElementData>thing)._updated! > (<ElementData>thingItMayReplace)._updated!) return true;
-        return false;
-    }
-
-    static hasSameId(a: ElementData, b: ElementData): boolean {
-        if (a._uid === b._uid) return true;
-        const type = Element.getTypeOfElement(a);
-        const typeB = Element.getTypeOfElement(b);
-        if (type !== typeB) return false;
-
-        if (type === 'DefData') return a._did === b._did;
-        if (type === 'EntryData') return a._eid === b._eid;
-        throw new Error('type not matched!')
-    }
-
-    private static getTypeOfElement(data: ElementData) {
-        if (data.hasOwnProperty("_eid")) return "EntryData"
-        if (data.hasOwnProperty("_did")) return "DefData"
-        return null
-    }
-
-    public static getMostRecent(elementArr: ElementData[]): ElementData | undefined {
-        if (elementArr.length == 0) {
-            console.warn('You tried to get the most recent of a list of zero elements');
-            return undefined
-        }
-        let mostRecent: ElementData = {
-            _did: '',
-            _uid: '',
-            _deleted: true,
-            _created: '',
-            _updated: '0' // earlier than anything before 12/31/1969, 6:00:00 PM CST
-        }
-        elementArr.forEach(element => {
-            if (element._updated! > mostRecent._updated!) mostRecent = element
-        })
-        return mostRecent
-    }
-}
-
-export class Def extends Element {
-    declare data: DefData;
-    private readonly _pts: PointDef[];
-
-    constructor(defData: DefLike, pdwRef: PDW) {
-        if (defData._scope !== undefined && !Def.isValidScope(defData._scope)) throw new Error('Invalid scope supplied when creating Def: ' + defData._scope);
-        super(defData, pdwRef);
-        this.data = {
-            ...this.data,
-            _lbl: defData._lbl ?? defData._did,
-            _desc: defData._desc ?? 'Set a description',
-            _emoji: defData._emoji ?? '🆕',
-            _scope: defData._scope ?? Scope.SECOND,
-            _tags: defData._tags ?? [],
-            _pts: []
-        }
-
-        if (this.data._lbl === 'Unlabeled Definition') this.data._lbl = 'Unlabeled Definition ' + this.data._did;
-
-        let pointsToSanitize = defData._pts ?? [];
-        //spawn new PointDefs for any non-underscore-prefixed keys
-        let pids = Object.keys(defData).filter(key => key.substring(0, 1) !== '_');
-        pids.forEach(pid => {
-            defData[pid]._pid = pid;
-            pointsToSanitize.push(defData[pid]);
-        })
-
-        this._pts = pointsToSanitize.map(rawPoint => new PointDef(rawPoint, this));
-        this.data._pts = this.pts.map(pd => pd.toData())
-
-        if (!Def.isDefData(this.data)) throw new Error('Def was mal-formed.');
-        // if (!this.data._deleted) 
-    }
-    get did() {
-        return this.data._did;
-    }
-    get lbl() {
-        return this.data._lbl;
-    }
-    get desc() {
-        return this.data._desc;
-    }
-    get emoji() {
-        return this.data._emoji;
-    }
-    get scope() {
-        return this.data._scope;
-    }
-    get tags() {
-        return this.data._tags;
-    }
-    public get pts(): PointDef[] {
-        return this._pts;
-    }
-
-    set lbl(newLbl: string) {
-        this.data._lbl = newLbl;
-        this.__modified = true;
-    }
-    set desc(newDesc: string) {
-        this.data._desc = newDesc;
-        this.__modified = true;
-    }
-    set emoji(newEmoji: string) {
-        if (!Def.isSingleEmoji(newEmoji)) {
-            console.warn('Tried to set an emoji with something that was too long, emoji will not change');
-            return
-        }
-        this.data._emoji = newEmoji;
-        this.__modified = true;
-    }
-
-    getPoint(pidOrLbl: string): PointDef | undefined {
-        let point = this.pts.find(point => point.pid === pidOrLbl);
-        if (point === undefined) point = this.pts.find(point => point.lbl.toUpperCase() === pidOrLbl.toUpperCase());
-        if (point !== undefined) return point
-        //warning maybe?
-        return undefined;
-        throw new Error('No point found with pid or lbl when getting "' + pidOrLbl + '" for the Def labeled "' + this.lbl + '"');
-    }
-
-    addPoint(pointInfo: PointDefLike): Def {
-        let newPoint = new PointDef(pointInfo, this);
-        this.pts.push(newPoint);
-        this.data._pts.push(newPoint.data)
-        this.__modified = true;
-        return this
-    }
-
-    /**
-     * Will return true of the Def has that tag, or the def has a tag in the passed-in array
-     */
-    hasTag(tag: string[] | string): boolean {
-        if (Array.isArray(tag)) return tag.some(item => this.data._tags.includes(item))
-        return this.data._tags.some(t => t.toUpperCase() === (<string>tag).toUpperCase());
-    }
-
-    addTag(tag: string): Def {
-        if (this.data._tags.some(t => t === tag)) {
-            console.warn('tried to add already-existing tag to def, ingoring request')
-        } else {
-            this.data._tags.push(tag);
-        }
-        return this
-    }
-
-    clearTags(): Def {
-        this.data._tags = [];
-        return this;
-    }
-
-    removeTag(tag: string): Def {
-        if (this.data._tags.some(t => t === tag)) {
-            this.data._tags = this.data._tags.filter(t => t !== tag);
-        } else {
-            console.warn('tried to remove a non-existant tag from a def, ingoring request')
-        }
-        return this
-    }
-
-    async newEntry(entryData: Omit<EntryLike, '_did'>): Promise<Entry> {
-        entryData._did = this.did;
-        return this.pdw.newEntry((<EntryLike>entryData));
-    }
-
-    /**
-    * Predicate to check if an object has all {@link DefData} properties
-    * AND they are the right type.
-    * @param data data to check
-    * @returns true if data have all required properties of {@link DefData}
-    */
-    static isDefData(data: any): boolean {
-        if (typeof data._did !== 'string') return false
-        if (typeof data._lbl !== 'string') return false
-        if (typeof data._desc !== 'string') return false
-        if (typeof data._emoji !== 'string') return false
-        if (!Def.isSingleEmoji(data._emoji)) return false
-        if (data._scope == undefined || !Def.isValidScope(data._scope)) return false
-        if (typeof data._uid !== 'string') return false
-        if (typeof data._created !== 'string') return false
-        if (typeof data._deleted !== 'boolean') return false
-        if (typeof data._updated !== 'string') return false
-        if (!Array.isArray(data._pts)) return false
-        if (!Array.isArray(data._tags)) return false
-        return true;
-    }
-
-    static isValidScope(typeStr: string): boolean {
-        const values = Object.values(Scope);
-        return values.includes(typeStr as unknown as Scope)
-    }
-
-    private static isSingleEmoji(emojiStr: string): boolean {
-        return [...new Intl.Segmenter().segment(emojiStr)].length == 1;
-    }
-}
-
-export class PointDef {
-    data: PointDefData
-    private _def: Def;
-    constructor(newPointDefData: PointDefLike, def: Def) {
-        if (newPointDefData._pid === undefined) throw new Error("No PointDef _pid supplied.");
-        if (newPointDefData._type !== undefined && !PointDef.isValidType(newPointDefData._type)) throw new Error('Cannot parse point type ' + newPointDefData._type);
-        if (newPointDefData._rollup !== undefined && !PointDef.isValidRollup(newPointDefData._rollup)) throw new Error('Cannot parse point rollup ' + newPointDefData._rollup);
-
-        //Force array for types that should have an array
-        if (
-            newPointDefData._opts === undefined &&
-            PointDef.shouldHaveOpts(newPointDefData._type!)
-        ) {
-            newPointDefData._opts = [];
-        }
-
-        this._def = def;
-
-        this.data = {
-            _pid: newPointDefData._pid,
-            _lbl: newPointDefData._lbl ?? newPointDefData._pid,
-            _type: newPointDefData._type ?? PointType.TEXT,
-            _desc: newPointDefData._desc ?? 'Set a description',
-            _emoji: newPointDefData._emoji ?? '🆕',
-            _rollup: newPointDefData._rollup ?? Rollup.COUNT,
-            _opts: newPointDefData._opts
-        }
-
-        if (!PointDef.isPointDefData(this.data)) throw new Error('Mal-formed PointDef')
-    }
-    get pid() {
-        return this.data._pid;
-    }
-    get lbl() {
-        return this.data._lbl;
-    }
-    get type() {
-        return this.data._type;
-    }
-    get desc() {
-        return this.data._desc;
-    }
-    get emoji() {
-        return this.data._emoji;
-    }
-    get opts() {
-        //@ts-expect-error /// somewhere I have something that forces data._opts to be defined
-        return this.data._opts;
-    }
-    get rollup() {
-        return this.data._rollup
-    }
-    public get def(): Def {
-        return this._def;
-    }
-
-    set lbl(newLbl: string) {
-        this.data._lbl = newLbl
-        this.def.__modified = true;
-    }
-    set desc(newDesc: string) {
-        this.data._desc = newDesc
-        this.def.data._pts.find(dp => dp._pid === this.data._pid)!._desc = newDesc;
-        // this.def.getPoint(this.data._pid).data._desc = newDesc;
-        this.def.__modified = true;
-    }
-    set emoji(newEmoji: string) {
-        if (!PointDef.isSingleEmoji(newEmoji)) {
-            console.warn('Tried to set an emoji with something that was too long, emoji will not change');
-            return
-        }
-        this.data._emoji = newEmoji
-        this.def.getPoint(this.data._pid)!.data._emoji = newEmoji;
-        this.def.__modified = true;
-    }
-    set opts(newOptArray: string[]) {
-        this.data._opts = newOptArray;
-        this.def.__modified = true;
-    }
-    set rollup(newRollup: Rollup) {
-        //#TODO - check if rollup is compatible with this.type
-        if (PointDef.isValidRollup(newRollup)) this.data._rollup = newRollup;
-    }
-
-    toData(): PointDefData {
-        return JSON.parse(JSON.stringify(this.data));
-    }
-
-    /**
-     * Passes along the save to the Def that owns this.
-     */
-    save() {
-        this.def.save();
-    }
-
-    hasOpt(lbl: string): boolean {
-        if (this.data._opts === undefined) return false;
-        const trimmedUpperCasedLabel = lbl.toUpperCase().trim();
-        return this.data._opts.some(opt => opt.toUpperCase().trim() === trimmedUpperCasedLabel);
-    }
-
-    addOpt(lbl: string) {
-        if (!PointDef.shouldHaveOpts(this.type)) {
-            console.warn('Cannot add option to PointDef of type ' + this.type);
-            return
-        }
-        const upperTrimmedLabel = lbl.toUpperCase().trim()
-        if (this.opts?.some(opt => opt.toUpperCase().trim() === upperTrimmedLabel)) {
-            console.warn('Tried adding an already existing option to PointDef ' + this.lbl + ', the option was: ' + lbl);
-            return
-        }
-        this.opts?.push(lbl);
-        this.def.__modified = true;
-    }
-
-    removeOpt(lbl: string) {
-        if (!PointDef.shouldHaveOpts(this.type)) {
-            console.warn('Cannot remove option for PointDef of type ' + this.type);
-            return
-        }
-        const upperTrimmedLabel = lbl.toUpperCase().trim()
-        if (!this.opts?.some(opt => opt.toUpperCase().trim() === upperTrimmedLabel)) {
-            console.warn('Tried removing an non-existant option from PointDef ' + this.lbl + ', the option was: ' + lbl);
-            return
-        }
-        this.opts = this.opts.filter(opt => opt.toUpperCase().trim() !== upperTrimmedLabel);
-        this.def.__modified = true;
-        return this;
-    }
-
-    /**
-     * The Def that contains this PointDef.
-     */
-    getDef(): Def {
-        return this.def;
-    };
-
-    shouldHaveOpts(): boolean {
-        return PointDef.shouldHaveOpts(this.type);
-    }
-
-    static shouldHaveOpts(type: string): boolean {
-        return type === PointType.SELECT || type === PointType.MULTISELECT;
-    }
-
-    /**
-    * Predicate to check if an object has all {@link PointDefData} properties
-    * AND they are the right type.
-    * @param data data to check
-    * @returns true if data have all required properties of {@link DefData}
-    */
-    static isPointDefData(data: any): boolean {
-        if (typeof data !== 'object') return false;
-        if (typeof data._lbl !== 'string') return false
-        if (typeof data._desc !== 'string') return false
-        if (typeof data._emoji !== 'string') return false
-        if (!PointDef.isSingleEmoji(data._emoji)) return false
-        if (data._type == undefined || !PointDef.isValidType(data._type)) return false
-        if (data._rollup == undefined || !PointDef.isValidRollup(data._rollup)) return false
-        return true;
-    }
-
-    /**
-    * Attempts to intercept poorly-typed values and convert them if possible.
-    * If conversion isn't possible, returns undefined and logs to console.warn
-    * @param _val value to check
-    * @param _type type to attempt to convert to
-    * @returns correctly-typed _val, or undefined if _val cannot be converted
-    */
-    ensureValType(_val: string | number | boolean | object, _type?: PointType) {
-        if (_type === undefined) _type = this.type;
-
-        if (_type === PointType.BOOL) {
-            if (typeof _val === 'boolean') return _val;
-            if (typeof _val === 'string') return _val.toUpperCase() === 'TRUE';
-            if (typeof _val === "number") return _val !== 0;
-            throw new Error(`Cannot convert this to boolean:` + _val)
-        }
-        if (_type === PointType.DURATION) {
-            if (typeof _val === 'string') return _val;
-            //if (typeof _val === "number") return _val !== 0; //number of seconds? millseconds?
-            if (typeof _val === 'number') {
-                // console.warn('Assuming excel portion of day for duration: ', _val)
-                return Temporal.Duration.from({ seconds: Math.round(86400 * _val) }).toString()
-            }
-            throw new Error(`Cannot convert this to duration:` + _val)
-        }
-        if (_type === PointType.MARKDOWN) {
-            if (typeof _val === 'boolean') return _val.toString();
-            if (typeof _val === 'string') return _val;
-            if (typeof _val === "number") return _val.toString();
-            throw new Error(`Cannot convert this to markdown:` + _val)
-        }
-        if (_type === PointType.MULTISELECT) {
-            if (Array.isArray(_val)) return _val;
-            if (typeof _val === 'string') return _val.split(',').map(text => text.trim());
-            throw new Error(`Cannot convert this to multiselect:` + _val)
-        }
-        if (_type === PointType.NUMBER) {
-            if (typeof _val === 'boolean') return _val ? 1 : 0;
-            if (typeof _val === 'string') return Number.parseInt(_val);
-            if (typeof _val === "number") return _val;
-            throw new Error(`Cannot convert this to number:` + _val)
-        }
-        if (_type === PointType.SELECT) {
-            if (typeof _val === 'string') return _val;
-            if (Array.isArray(_val)) return _val.join('|||')
-            if (_val === undefined) return []; //I guess
-            throw new Error(`Cannot convert this to select:` + _val);
-        }
-        if (_type === PointType.TEXT) {
-            if (typeof _val === 'boolean') return _val.toString();
-            if (typeof _val === 'string') return _val.trim();
-            if (typeof _val === "number") return _val.toString();
-            if (_val === undefined) return ''; //I guess
-            return _val.toString();
-        }
-        if (_type === PointType.TIME) {
-            if (typeof _val === 'string') return _val.trim();
-            if (typeof _val === "number" && _val >= 0 && _val <= 1) {
-                //attempt to support "proportion of day to time" like Excel would
-                return new Temporal.PlainTime(0, 0).add({ seconds: Math.round(86400 * _val) }).toString();
-            }
-            throw new Error(`Cannot convert this to time:` + _val)
-        }
-        throw new Error("Type not supported: " + _type);
-    }
-
-    static isValidType(typeStr: string): boolean {
-        //Handy bit of Enum functionality here for ref
-        const values = Object.values(PointType);
-        return values.includes(typeStr as unknown as PointType)
-    }
-
-    static isValidRollup(typeStr: string): boolean {
-        //Handy bit of Enum functionality here for ref
-        const values = Object.values(Rollup);
-        return values.includes(typeStr as unknown as Rollup)
-    }
-
-    private static isSingleEmoji(emojiStr: string): boolean {
-        return [...new Intl.Segmenter().segment(emojiStr)].length == 1;
-    }
-}
-
-export class Entry extends Element {
-    declare data: EntryData;
-    private _def: Def;
-    private _period: Period;
-    /**
-     * Creates a LOCAL entry instance
-     * @param entryData data to use to build the entry
-     * @param def def to use, if none is supplied it will pull from the local manifest based on entryData._did
-     */
-    constructor(entryData: EntryLike, pdwRef: PDW) {
-        super(entryData, pdwRef);
-        let def = pdwRef.getDefFromManifest(entryData._did);
-        if (def === undefined) throw new Error("Tried creating an Entry with _did '" + entryData._did + "', which is not in the manifest.");
-        this._def = def;
-
-        //forcing scope adherance
-        if (entryData._period !== undefined) {
-            this._period = new Period(entryData._period, this.def.scope);
-        } else {
-            this._period = Period.nowAsPeriod(this.def.scope);
-        }
-
-        this.data = {
-            ...this.data,
-            _period: this.period.toString(),
-            _eid: entryData._eid ?? makeUID(),
-            _note: entryData._note ?? '',
-            _source: entryData._source ?? ''
-        }
-
-        //migrate points up - to support using a _pts: map on the entryData object. Not preferred, but may be handy.
-        //probably YAGNI
-        if (entryData.hasOwnProperty('_pts')) {
-            Object.keys(entryData._pts).forEach(key => {
-                entryData[key] = entryData._pts[key];
-            })
-        }
-
-        //spawn new EntryPoints for any non-underscore-prefixed keys
-        let pids = Object.keys(entryData).filter(key => key.substring(0, 1) !== '_');
-        pids.forEach(pid => {
-            const keyval = this.sanitizePointData({ [pid]: entryData[pid] });
-            this.data[keyval.pid] = keyval.val;
-        })
-
-        if (!Entry.isEntryData(this.data)) throw new Error('The Entry was mal-formed');
-    }
-
-    get did() {
-        return this.data._did;
-    }
-    get eid() {
-        return this.data._eid;
-    }
-    public get period(): Period {
-        return this._period;
-    }
-    get note() {
-        return this.data._note;
-    }
-    get source() {
-        return this.data._source;
-    }
-    public get def() {
-        return this._def;
-    }
-
-    set note(newNote: string) {
-        this.data._note = newNote;
-        this.__modified = true;
-    }
-    set source(newSource: string) {
-        this.data._source = newSource;
-        this.__modified = true;
-    }
-    public set period(value: Period | string) {
-        if (typeof value === 'string') value = new Period(value);
-        //ensure scope agreement
-        value = value.zoomTo(this.def.scope)
-        this.data._period = value.toString();
-        this._period = value;
-        this.__modified = true;
-    }
-
-    /**
-     * Looks up the associated PointDef based on the object key.
-     * Checks the value against that PointDef._type. Calls {@link PointDef.ensureValType()}
-     * @param rawPointData key: value pair with key = _pid or _lbl of PointDef
-     * @returns an object with 'key' = _pid for PointDef, and 'val' = value
-     */
-    sanitizePointData(rawPointData: any): { 'pid': string, 'lbl': string, 'val': any, 'pointDef': PointDef } {
-        if (typeof rawPointData !== 'object') throw new Error('Saw non-object Entry Point:' + rawPointData);
-        if (Object.keys(rawPointData).length > 1) throw new Error('Expected a single-keyed object. Saw multiple keys');
-        const key = Object.keys(rawPointData)[0];
-
-        const assPointDef = this.def?.getPoint(key);
-        if (assPointDef === undefined) throw new Error('No PointDef found for ' + key)
-        const validatedValue = assPointDef?.ensureValType(rawPointData[key]);
-
-        return {
-            'pid': assPointDef.pid,
-            'lbl': assPointDef.lbl,
-            'val': validatedValue,
-            'pointDef': assPointDef
-        }
-    }
-
-    getPoints(): { 'pid': string, 'lbl': string, 'val': any, 'pointDef': PointDef }[] {
-        let pids = Object.keys(this.data).filter(key => key.substring(0, 1) !== '_');
-        return pids.map(pid => {
-            const assPD = this.def.pts.find(point => point.pid === pid);
-            return {
-                pointDef: assPD!,
-                pid: pid,
-                lbl: assPD!.lbl,
-                val: this.data[pid]
-            }
-        })
-    }
-
-    getPoint(pidOrLbl: string): { 'pid': string, 'lbl': string, 'val': any, 'pointDef': PointDef } | undefined {
-        const allPoints = this.getPoints();
-        let point = allPoints.find(point => point.pid === pidOrLbl);
-        if (point !== undefined) return point;
-        return allPoints.find(point => point.lbl === pidOrLbl);
-    }
-
-    getPointVal(pidOrLbl: string): any {
-        return this.getPoint(pidOrLbl)!.val;
-    }
-
-    setPointVals(pidValArr: { [pid: string]: any }[]): Entry {
-        pidValArr.forEach(keyVal => {
-            let pointVal = this.sanitizePointData(keyVal);
-            this.data[pointVal.pid] = pointVal.val;
-        })
-        this.__modified = true;
-        return this;
-    }
-
-    setPointVal(pid: string, val: any): Entry {
-        this.__modified = true;
-        return this.setPointVals([{ [pid]: val }]);
-    }
-
-    /**
-    * Predicate to check if an object has all {@link EntryData} properties
-    * AND they are the right type. Does not check for EntryPoints or their types.
-    * @param data data to check
-    * @returns true if data have all required properties of {@link DefData}
-    */
-    static isEntryData(data: any): boolean {
-        if (typeof data._did !== 'string') return false
-        if (typeof data._eid !== 'string') return false
-        if (typeof data._note !== 'string') return false
-        if (typeof data._period !== 'string') return false
-        if (typeof data._uid !== 'string') return false
-        if (typeof data._created !== 'string') return false
-        if (typeof data._deleted !== 'boolean') return false
-        if (typeof data._updated !== 'string') return false
-        if (typeof data._source !== 'string') return false
-        return true;
-    }
-}
-
-/**
- * Periods are Immutable. All methods return new copies of Period.
- * They are basically wrappers around a string.
- */
-export class Period {
-    /**
-     * The Period itself, represented in a {@link PeriodStr} format.  
-     * 
-     * **Examples**:
-     * - '2020'
-     * - '2020-Q1'
-     * - '2020-03'
-     * - '2020-03-19'
-     * - '2020-03-19T18'
-     * - '2020-03-19T18:59'
-     * - '2020-03-19T18:59:25'
-     */
-    periodStr: PeriodStr;
-    scope: Scope;
-    private _zoomLevel: number
-
-    constructor(periodStr: PeriodStr | Period, desiredScope?: Scope) {
-        if (typeof periodStr !== 'string') periodStr = periodStr.periodStr;
-        this.periodStr = periodStr;
-        this.scope = Period.inferScope(periodStr);
-        if (this.scope === Scope.SECOND && this.periodStr.length > 20) this.periodStr = convertFullISOToPlainDateStr(this.periodStr);
-        this._zoomLevel = Period.zoomLevel(this.scope)
-
-        if (desiredScope !== undefined && this.scope !== desiredScope) {
-            console.log('Converting ' + periodStr + ' to scope ' + desiredScope);
-            return this.zoomTo(desiredScope);
-        }
-
-        /**
-         * This sucked to figure out and was annoying to need to do at all.
-         * @param isoString an ISO formatting string, along with the timezone offset
-         * @returns an ISO formatted string without the offset, representing the time *in that timezone*
-         */
-        function convertFullISOToPlainDateStr(isoString: string): string {
-            let temp = Temporal.PlainDateTime.from(isoString);
-            const tzString = Temporal.TimeZone.from(isoString).toString();
-            const offsetString = tzString.substring(0, 1) + 'PT' + Number.parseInt(tzString.substring(1, 3)) + 'h' + + Number.parseInt(tzString.substring(5, 7)) + 'm';
-            console.log(offsetString);
-            temp = temp.add(offsetString);
-            let str = temp.toJSON()
-            return str
-        }
-    }
-
-    private static zoomLevel(scope: Scope): number {
-        return [
-            Scope.SECOND,
-            Scope.MINUTE,
-            Scope.HOUR,
-            Scope.DAY,
-            Scope.WEEK,
-            Scope.MONTH,
-            Scope.QUARTER,
-            Scope.YEAR
-        ].findIndex(val => val === scope)
-    }
-
-    /**
-     * Yay overriding default Object prototype methods!
-     * @returns periodStr
-     */
-    toString() {
-        return this.periodStr;
-    }
-
-    toTemporalPlainDate() {
-        return Temporal.PlainDate.from(this.periodStr);
-    }
-
-    toTemporalPlainDateTime() {
-        return Temporal.PlainDateTime.from(this.periodStr);
-    }
-
-    getEnd(): Period {
-        if (this.scope === Scope.SECOND) return new Period(this.periodStr);
-        if (this.scope === Scope.MINUTE) return new Period(this.periodStr + ':59');
-        if (this.scope === Scope.HOUR) return new Period(this.periodStr + ':59:59');
-        if (this.scope === Scope.DAY) return new Period(this.periodStr + 'T23:59:59');
-
-        if (this.scope === Scope.WEEK) {
-            let numWks = Number.parseInt(this.periodStr.split('W')[1]) - 1;
-            //some years dont' start until after the 1st
-            if (Period.needsWeekShift(this.periodStr.substring(0, 4))) numWks = numWks + 1
-            let init = Temporal.PlainDate.from(this.periodStr.split('-')[0] + '01-01')
-
-            let sun = init.add({ days: 7 - init.dayOfWeek })
-            sun = sun.add({ days: numWks * 7 });
-            return new Period(sun.toString() + 'T23:59:59')
-
-        }
-        if (this.scope === Scope.MONTH) {
-            let lastDay = Temporal.PlainDate.from(this.periodStr + '-01').daysInMonth;
-            return new Period(this.periodStr + '-' + lastDay.toString() + 'T23:59:59')
-        }
-        if (this.scope === Scope.QUARTER) {
-            const year = this.periodStr.substring(0, 4)
-            const q = Number.parseInt(this.periodStr.slice(-1));
-            const month = q * 3
-            const d = Temporal.PlainDate.from(year + '-' + month.toString().padStart(2, '0') + '-01').daysInMonth;
-            return new Period(year + '-' + month.toString().padStart(2, '0') + '-' + d + 'T23:59:59');
-        }
-        return new Period(this.periodStr + "-12-31T23:59:59")
-    }
-
-    /**
-     * 
-     * @returns the first second of the period (e.g. 2020-01-01T00:00:00)
-     */
-    getStart(): Period {
-        if (this._zoomLevel === 0) return new Period(this);
-        if (this.scope === Scope.YEAR) return new Period(this.toString() + '-01-01T00:00:00')
-        if (this.scope === Scope.QUARTER) return new Period(this.zoomIn() + '-01T00:00:00')
-        if (this.scope === Scope.MONTH) return new Period(this.toString() + '-01T00:00:00')
-        //above preempts week, cause it's not purely hierarchical,
-        //from here you can just "zoomIn" to the beginning of the period
-        let per = this.zoomIn();
-        while (per._zoomLevel !== 0) {
-            per = per.zoomIn()
-        }
-        return per;
-    }
-
-    /**
-     * 
-     */
-    zoomTo(desiredScope: Scope): Period {
-        const desiredLevel = Period.zoomLevel(desiredScope);
-        if (this._zoomLevel === desiredLevel) return new Period(this);
-        if (this._zoomLevel < desiredLevel) {
-            let zoomOut = this.zoomOut()
-            while (zoomOut._zoomLevel < desiredLevel) {
-                //need to bypass weeks
-                if (desiredLevel !== 4 && zoomOut._zoomLevel === 3) {
-                    zoomOut = new Period(zoomOut.periodStr.substring(0, 7))
-                } else {
-                    zoomOut = zoomOut.zoomOut()
-                }
-            }
-            return zoomOut;
-        }
-        let zoomIn = this.zoomIn()
-        while (zoomIn._zoomLevel > desiredLevel) {
-            if (desiredLevel !== 4 && zoomIn._zoomLevel === 5) {
-                zoomIn = new Period(zoomIn.periodStr + '-01')
-            } else {
-                zoomIn = zoomIn.zoomIn()
-            }
-        }
-        return zoomIn;
-    }
-
-    /**
-     * Zooms in on the BEGINNING (?) of the Period
-     * @returns the next level finer-grain scope at the beginning of this scope
-     */
-    zoomIn(): Period {
-        if (this.scope === Scope.YEAR) return new Period(this.periodStr + '-Q1');
-        if (this.scope === Scope.QUARTER) {
-            const year = this.periodStr.substring(0, 4);
-            const month = Number.parseInt(this.periodStr.slice(-1)) * 3 - 2
-            return new Period(year + '-' + month.toString().padStart(2, '0'));
-        }
-        if (this.scope === Scope.MONTH) {
-            const temp = Temporal.PlainDate.from(this.periodStr + "-01")
-            let year = temp.year;
-            if (temp.weekOfYear > 50 && temp.dayOfYear < 14) year = year - 1;
-            if (temp.weekOfYear == 1 && temp.dayOfYear > 360) year = year + 1
-            const weekNum = temp.weekOfYear
-            return new Period(year + "-W" + weekNum.toString().padStart(2, '0'));
-        }
-        if (this.scope === Scope.WEEK) {
-            let numWks = Number.parseInt(this.periodStr.split('W')[1]) - 1;
-            //if the previous year had 53 weeks, this is necessary
-            if (Period.needsWeekShift(this.periodStr.substring(0, 4))) numWks = numWks + 1
-            let init = Temporal.PlainDate.from(this.periodStr.split('-')[0] + '01-01')
-            let mon = init.add({ days: 1 - init.dayOfWeek }).add({ days: numWks * 7 })
-            return new Period(mon.toString());
-        }
-        if (this.scope === Scope.DAY) return new Period(this.periodStr + "T00")
-        if (this.scope === Scope.HOUR) return new Period(this.periodStr + ":00")
-        if (this.scope === Scope.MINUTE) return new Period(this.periodStr + ":00")
-        //zooming in from a second returns itself
-        return new Period(this.periodStr)
-    }
-
-    zoomOut(): Period {
-        if (this.scope === Scope.SECOND) return new Period(this.periodStr.slice(0, -3));
-        if (this.scope === Scope.MINUTE) return new Period(this.periodStr.slice(0, -3));
-        if (this.scope === Scope.HOUR) return new Period(this.periodStr.slice(0, -3));
-        if (this.scope === Scope.DAY) {
-            //this SHOULD be usign Temporal.PlaidDateTime.weekOfYear(), but that's not implemented on this polyfill
-            const temp = Temporal.PlainDateTime.from(this.periodStr);
-            //catching edge cases like 2019-12-31 => 2020-W01 & 2023-01-01 => 2022-W52
-            let year = temp.year;
-            if (temp.weekOfYear > 50 && temp.dayOfYear < 14) year = year - 1;
-            if (temp.weekOfYear == 1 && temp.dayOfYear > 360) year = year + 1
-            return new Period(year + "-W" + temp.weekOfYear.toString().padStart(2, '0'));
-        }
-        if (this.scope === Scope.WEEK) {
-            return new Period(Period.getMidWeek(this).toString().substring(0, 7))
-            // //weeks zooming out resolve to whichever month contains the THURSDAY of the week
-            // let numWks = Number.parseInt(this.periodStr.split('W')[1]) - 1;
-            // //if the previous year had 53 weeks, this is necessary
-            // if (Period.prevYearHas53Weeks(this.periodStr.substring(0, 4))) numWks = numWks + 1
-            // let init = Temporal.PlainDate.from(this.periodStr.split('-')[0] + '01-01')
-            // let thur = init.add({ days: 4 - init.dayOfWeek }).add({ days: numWks * 7 })
-            // return new Period(thur.toPlainYearMonth().toString());
-        }
-        if (this.scope === Scope.MONTH) {
-            let yearStr = this.periodStr.split('-')[0];
-            let month = Number.parseInt(this.periodStr.split('-')[1]);
-            let quarterStr = Math.ceil(month / 3).toString();
-            return new Period(yearStr + '-Q' + quarterStr);
-        }
-        // else is a Scope.QUARTER or Scope.YEAR, I think I'm goign to let Scope.YEAR return itself
-        return new Period(this.periodStr.substring(0, 4))
-    }
-
-    addDuration(temporalDurationStr: string): Period {
-        const startTemp = Temporal.PlainDateTime.from(this.getStart().periodStr);
-        const endTemp = startTemp.add(temporalDurationStr);
-        return new Period(endTemp.toString()).zoomTo(this.scope);
-    }
-
-    contains(period: Period | PeriodStr): boolean {
-        if (typeof period === 'string') period = new Period(period);
-        //converting week scopes to THURSDAY of that week
-        if (period.scope === Scope.WEEK) period = Period.getMidWeek(period)
-        const inBegin = Temporal.PlainDateTime.from(period.getStart().periodStr)
-        const inEnd = Temporal.PlainDateTime.from(period.getEnd().periodStr)
-        const thisBegin = Temporal.PlainDateTime.from(this.getStart().periodStr);
-        const thisEnd = Temporal.PlainDateTime.from(this.getEnd().periodStr);
-        const start = Temporal.PlainDateTime.compare(inBegin, thisBegin);
-        const end = Temporal.PlainDateTime.compare(thisEnd, inEnd);
-        return start !== -1 && end !== -1
-    }
-
-    isBefore(period: Period): boolean {
-        const start = Temporal.PlainDateTime.from(this.getEnd().periodStr);
-        const end = Temporal.PlainDateTime.from(period.getStart().periodStr);
-        return Temporal.PlainDateTime.compare(start, end) === -1
-    }
-
-    isAfter(period: Period): boolean {
-        const start = Temporal.PlainDateTime.from(this.getStart().periodStr);
-        const end = Temporal.PlainDateTime.from(period.getEnd().periodStr);
-        return Temporal.PlainDateTime.compare(start, end) === 1
-    }
-
-    // I can't believe I was able to reduce these to a 1 liner
-    getNext(): Period {
-        // let start = this.getStart();
-        // let end = this.getEnd();
-        // let plusOne = end.addDuration('PT1S');
-        // let zoomed = plusOne.zoomTo(this.scope);
-        return this.getEnd().addDuration('PT1S').zoomTo(this.scope);
-
-    }
-    getPrev(): Period {
-        return this.getStart().addDuration('-PT1S').zoomTo(this.scope);
-    }
-
-    private static needsWeekShift(yearStr: string): boolean {
-        return Temporal.PlainDate.from(yearStr + '-01-01').dayOfWeek > 4;
-
-    }
-
-    private static getMidWeek(period: Period) {
-        if (period.scope !== Scope.WEEK) return period;
-        return period.getStart().zoomTo(Scope.DAY).addDuration('P3D');
-    }
-
-    static allPeriodsIn(start: Period, end: Period, scope: Scope, asStrings = false): Period[] | string[] {
-        if (Temporal.PlainDateTime.compare(Temporal.PlainDateTime.from(start.getStart().periodStr), Temporal.PlainDateTime.from(end.getStart().periodStr)) === 1) {
-            console.warn('You may have flipped your start and end dates accidentally... or something')
-            const temp = start;
-            start = end;
-            end = temp;
-        }
-        // const startOfStart = start.getStart().periodStr;
-        // const endOfEnd = end.getEnd().periodStr;
-        let first, last, list: any[];
-        first = start.zoomTo(scope);
-        last = end.getNext().zoomTo(Scope.SECOND).addDuration('-PT1S').zoomTo(scope);
-
-        let member = first;
-        list = [];
-        if (asStrings) {
-            do {
-                // list.push({[member.periodStr]: {from: member.getStart().toString(), to: member.getEnd().toString()}});
-                list.push(member.periodStr);
-                member = member.getNext();
-            } while (member.periodStr <= last.periodStr)
-            return list as string[]
-        } else {
-            do {
-                list.push(member);
-                member = member.getNext();
-
-            } while (member.periodStr <= last.periodStr)
-            return list as Period[]
-        }
-    }
-
-    //getEntriesInPeriodMatchingFilter(params: StandardFilters): Element[]
-
-    static now(scope: Scope): PeriodStr {
-        let seedStr = '';
-        let nowTemp = Temporal.Now.zonedDateTimeISO();
-        if (scope === Scope.YEAR) seedStr = nowTemp.year.toString();
-        if (scope === Scope.QUARTER) seedStr = nowTemp.year.toString() + '-Q' + Math.ceil(nowTemp.month / 3);
-        if (scope === Scope.MONTH) seedStr = nowTemp.toPlainYearMonth().toString()
-        if (scope === Scope.WEEK) seedStr = nowTemp.year.toString() + '-W' + nowTemp.weekOfYear.toString();
-        if (scope === Scope.DAY) seedStr = nowTemp.toPlainDate().toString()
-        if (scope === Scope.HOUR) seedStr = nowTemp.toString().substring(0, 13)
-        if (scope === Scope.MINUTE) seedStr = nowTemp.toString().substring(0, 16)
-        if (scope === Scope.SECOND) seedStr = nowTemp.toString().substring(0, 19)
-        return seedStr
-    }
-
-    static nowAsPeriod(scope: Scope): Period {
-        return new Period(Period.now(scope));
-    }
-
-    static inferScope(ISOString: string): Scope {
-        // /https://xkcd.com/208/
-        if (/\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d:[0-5]\d/i.test(ISOString))
-            return Scope.SECOND;
-        if (/\d{4}-[01]\d-[0-3]\dT[0-2]\d:[0-5]\d/i.test(ISOString))
-            return Scope.MINUTE;
-        if (/\d{4}-[01]\d-[0-3]\dT[0-2]\d/i.test(ISOString))
-            return Scope.HOUR;
-        if (/^([12]\d{3}-(0[1-9]|1[0-2])-(0[1-9]|[12]\d|3[01]))$/.test(ISOString))
-            return Scope.DAY;
-        if (/^([0-9]{4})-?W(5[0-3]|[1-4][0-9]|0[1-9])$/i.test(ISOString))
-            return Scope.WEEK;
-        if (/^([0-9]{4})-(1[0-2]|0[1-9])$/.test(ISOString))
-            return Scope.MONTH;
-        if (/^[0-9]{4}-Q[1-4]$/i.test(ISOString))
-            return Scope.QUARTER;
-        if (/^([0-9]{4})$/.test(ISOString))
-            return Scope.YEAR;
-        throw new Error('Attempted to infer scope failed for: ' + ISOString);
     }
 }
